Rename datafields and move stuff to EE folder.
