+++ Content.Shared/_EinsteinEngines/TelescopicBaton/KnockdownOnHitEvent.cs
@@ -1,6 +1,6 @@
 using Content.Shared._White.Standing;
 
-namespace Content.Shared.Stunnable.Events;
+namespace Content.Shared._EinsteinEngines.TelescopicBaton;
 
 [ByRefEvent]
 public record struct KnockdownOnHitAttemptEvent(bool Cancelled, DropHeldItemsBehavior Behavior); // Goob edit
