+++ Resources/Prototypes/_EinsteinEngines/Entities/Objects/Weapons/Melee/telescopic_baton.yml
